simplify browser capture
