+++ packages/atomic-a11y/a11y/a11y-manual-criteria.json
@@ -7,14 +7,11 @@
     "1.4.4-resize-text",
     "1.4.5-images-of-text",
     "1.4.10-reflow",
-    "1.4.11-non-text-contrast",
     "1.4.12-text-spacing",
     "1.4.13-content-on-hover-focus",
     "2.4.4-link-purpose",
     "2.4.6-headings-and-labels",
     "2.4.7-focus-visible",
-    "2.4.11-focus-not-obscured",
-    "2.5.8-target-size",
-    "3.3.3-error-suggestion"
+    "2.4.11-focus-not-obscured"
   ]
 }
