+++ packages/atomic-a11y/package.json
@@ -15,7 +15,6 @@
     "a11y:merge-shards": "pnpm build && node scripts/merge-shards-cli.mjs"
   },
   "dependencies": {
-    "exponential-backoff": "^3.1.1",
     "js-yaml": "^4.1.0",
     "openai": "^4.77.0",
     "playwright": "^1.49.1"
